start to restore temp disabled features

+++ Asset/Shaders/Metal/texture.frag.metal
@@ -50,7 +50,7 @@ struct LightInfo
 
 struct DebugConstants
 {
-    float layer_index;
+    int layer_index;
     float mip_level;
     float line_width;
     float padding0;
