+++ Framework/Common/cbuffer.h
@@ -76,7 +76,7 @@ unistruct LightInfo REGISTER(b12)
 
 unistruct DebugConstants REGISTER(b13)
 {
-	float layer_index;
+	int layer_index;
 	float mip_level;
 	float line_width;
 	float padding0;
